refactor: clean up intake module dead code and promote supabase client
